+++ src/resources/containers/container.ts
@@ -56,6 +56,7 @@ export interface Container extends Resource<ContainerMetas> {
   stateful: boolean;
   requirements?: string[];
   annotations: Record<string, any> | null;
+  deprecate: boolean;
   state: State<ContainerState> & {
     desired: ContainerState | "";
   };
@@ -142,6 +143,7 @@ export interface CreateParams {
   config: Config;
   annotations?: Record<string, any>;
   volumes: Volumes.Volume[];
+  deprecate: boolean;
 }
 
 export async function create(